Restructuring
Continued restructuring of package for orginazational purposes.
