+++ tests/calInfoTest.py
